add playlist list when edit

+++ src/server/api/routers/playlist.ts
@@ -1,9 +1,11 @@
+import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 import {
   createTRPCRouter,
   protectedProcedure,
   publicProcedure,
   playlistReadProcedure,
+  playlistEditProcedure,
 } from "~/server/api/trpc";
 
 export const playlistRouter = createTRPCRouter({
@@ -28,7 +30,6 @@ export const playlistRouter = createTRPCRouter({
             description,
             readPrivacy,
             writePrivacy,
-            trackIds: [],
             ownerId: userId,
           },
         });
@@ -65,14 +66,57 @@ export const playlistRouter = createTRPCRouter({
     return ctx.prisma.playlist.findMany({});
   }),
 
+  addTrack: playlistEditProcedure
+    .input(z.object({ trackId: z.string() }))
+    .mutation(async ({ input: { trackId }, ctx }) => {
+      const playlistTracks = await ctx.prisma.playlist.findUnique({
+        where: {
+          id: ctx.playlist.id,
+          AND: { tracks: { some: { spotifyId: trackId } } },
+        },
+      });
+      if (playlistTracks === null) {
+        const addTrack = await ctx.prisma.playlist.update({
+          where: {
+            id: ctx.playlist.id,
+          },
+          data: {
+            tracks: { create: { spotifyId: trackId } },
+          },
+        });
+        return addTrack;
+      }
+
+      throw new TRPCError({
+        code: "BAD_REQUEST",
+        message: "Song already exists in playlist.",
+      });
+    }),
   delete: protectedProcedure
     .input(z.object({ id: z.string() }))
     .mutation(async ({ input: { id }, ctx }) => {
-      const deletedPlaylist = await ctx.prisma.playlist.delete({
+      const playlistOwnerId = await ctx.prisma.playlist.findUnique({
         where: { id },
+        select: { ownerId: true },
       });
+      if (playlistOwnerId?.ownerId === ctx.session.user.id) {
+        const deletedPlaylist = await ctx.prisma.playlist.delete({
+          where: { id },
+        });
 
-      return deletedPlaylist;
+        return deletedPlaylist;
+      }
+
+      if (playlistOwnerId?.ownerId !== ctx.session.user.id)
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "You are not authorized to delete this playlist.",
+        });
+
+      throw new TRPCError({
+        code: "NOT_FOUND",
+        message: "The server cannot find the requested resource.",
+      });
     }),
   getPlaylist: playlistReadProcedure.query(({ ctx }) => {
     return {
